inventory done

+++ src/main/java/com/warehouse/admin/repository/ClientRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.warehouse.admin.entity.Client;
 
-public interface ClientRepo extends JpaRepository<Client, Integer>{
+public interface ClientRepo extends JpaRepository<Client, Long>{
  
 	 boolean existsByEmail(String email);
 	    Optional<Client> findByEmail(String email);
